Moved to new package and added new relation in Role (directly to Permission, without class between them)
